Framework for AST-based checkers with some basic naming checkers
In order to accurately do checks on specific parts of the python grammar,
it makes sense to use the tools built into the standard library in order
check the proper parts of the grammar.  This commit adds a framework
for doing those checks as well as some additional functionality built
on the new framework code for doing checks against PEP8's prescriptive
naming conventions.  Follow-on commits should clean up those checkers
and add additional checkers.